pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["Alexey Vorobyov <alexey.vorobyov.s@gmail.com>"]
description = "Library management system"
name = "django-libms"
package-mode = false
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
Django = "^5.0"
bpython = "^0.24"
django-cors-headers = "^4.4.0"
django-environ = "^0.11.2"
django-filter = "^24.2"
django-split-settings = "^1.3.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
python = "~3.11"

[tool.poetry.group.dev.dependencies]
mixer = "^7.2.2"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-deadfixtures = "^2.2.1"
pytest-django = "^4.8.0"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.14.0"
ruff = "^0.5.1"
toml-sort = "^0.23.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
addopts = [
    "--no-migrations",
    # https://pytest-django.readthedocs.io/en/latest/database.html#reuse-db-reuse-the-testing-database-between-test-runs
    "--reuse-db",
    "-p no:warnings",
    "-s",
    "-v",
]
python_files = "test*.py"
pythonpath = ". src"

[tool.ruff]
line-length = 160
target-version = "py311"

[tool.ruff.lint]
extend-select = [
    "TID252",  # Prefer absolute imports over relative imports from parent modules
]
ignore = []

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
known-local-folder = ["apps", "core"]

[tool.tomlsort]
all = true
in_place = true
sort_first = ["tool.poetry"]
spaces_before_inline_comment = 2
spaces_indent_inline_array = 4
trailing_comma_inline_array = true
